Added all tests for add handlers with some fixes

+++ StringManager/StringManager.Core/Models/ToneInTuning.cs
@@ -8,8 +8,8 @@
 
         public string ToneName { get; set; }
 
-        public int Frequency { get; set; }
+        public double Frequency { get; set; }
 
-        public int WaveLenght { get; set; }
+        public double WaveLenght { get; set; }
     }
 }
